refactor: PIZ-3 decorator pattern for json response
