+++ app/controllers/base.py
@@ -1,6 +1,6 @@
 from typing import Any, Optional, Tuple
 from sqlalchemy.exc import SQLAlchemyError
-from ..repositories.managers import BaseManager
+from ..repositories.managers.base import BaseManager
 
 
 class BaseController:
