+++ app/controllers/index.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 from sqlalchemy.exc import SQLAlchemyError
 
-from ..repositories.managers import IndexManager
+from ..repositories.managers.db_status import DBStatusManager
 
 
 class IndexController:
@@ -9,7 +9,7 @@ class IndexController:
     @staticmethod
     def test_connection() -> Tuple[bool, str]:
         try:
-            IndexManager.test_connection()
+            DBStatusManager.test_connection()
             return True, ''
         except (SQLAlchemyError, RuntimeError) as ex:
             return False, str(ex)
