+++ app/handlers/error_handlers.py
@@ -1,0 +1,9 @@
+
+
+def default_error_status_handler(data, error):
+    if data:
+        return 200
+    elif not error:
+        return 404
+    else:
+        return 400
