+++ app/repositories/managers/beverage.py
@@ -1,0 +1,14 @@
+from ..models import Beverage
+from ..serializers import BeverageSerializer
+from .base import BaseManager
+
+from typing import Sequence
+
+
+class BeverageManager(BaseManager):
+    model = Beverage
+    serializer = BeverageSerializer
+
+    @classmethod
+    def get_by_id_list(cls, ids: Sequence):
+        return cls.session.query(cls.model).filter(cls.model._id.in_(set(ids))).all() or []
