+++ app/repositories/serializers.py
@@ -33,6 +33,7 @@ class BeverageSerializer(ma.SQLAlchemyAutoSchema):
         model = Beverage
         load_instance = True
         fields = ('_id', 'name', 'price')
+
 class BeverageDetailSerializer(ma.SQLAlchemyAutoSchema):
 
     beverage = ma.Nested(BeverageSerializer)
@@ -44,6 +45,7 @@ class BeverageDetailSerializer(ma.SQLAlchemyAutoSchema):
             'beverage_price',
             'beverage'
         )
+        
 class OrderSerializer(ma.SQLAlchemyAutoSchema):
     size = ma.Nested(SizeSerializer)
     ingredients = ma.Nested(IngredientDetailSerializer, many=True)
