+++ app/services/ingredient.py
@@ -1,8 +1,8 @@
-from app.common.http_methods import GET, POST, PUT
-from flask import Blueprint
-from app.services.base import BaseService
-
+from ..common.http_methods import GET, POST, PUT
+from ..services.base import BaseService
 from ..controllers import IngredientController
+
+from flask import Blueprint
 
 ingredient = Blueprint('ingredient', __name__)
 baseService = BaseService(IngredientController)
