+++ app/services/report.py
@@ -1,13 +1,12 @@
-from app.common.http_methods import GET
-from flask import Blueprint, jsonify
-
+from ..common.http_methods import GET
+from ..decorators.handle_json_response import handle_json_response
 from ..controllers import ReportController
+
+from flask import Blueprint
 
 report = Blueprint('report', __name__)
 
+@handle_json_response
 @report.route('/', methods=GET)
-def get_report():
-    report, error = ReportController.get_reports()
-    response = report if not error else {'error': error}
-    status_code = 200 if not error else 400
-    return jsonify(response), status_code
+def get_reports():
+    return ReportController.get_reports()
