+++ seed/__init__.py
@@ -1,9 +1,9 @@
+from .generic_item_seeder import GenericItemSeeder
+from .custom_seeders import SizeSeeder, OrderSeeder
+from .handle_seeding import DBSeeder
+
 from app.controllers import BeverageController, IngredientController, SizeController, OrderController
 
-
-from seed.generic_item_seeder import GenericItemSeeder
-from seed.custom_seeders import SizeSeeder, OrderSeeder
-from seed.handle_seeding import DBSeeder
 
 db_seeder = DBSeeder()
 db_seeder.add_seeder(GenericItemSeeder(IngredientController, "ingredient"))
